requirements.txt
attrs==22.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pymongo==4.3.3
pyodbc==4.0.35
pytest==7.2.1
python-decouple==3.7
python-dotenv==0.21.1
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.12.1
